requirements.txt
python-telegram-bot==13.15
pytelegrambotapi==4.13.0
python-dotenv
google-cloud-aiplatform
vertexai